remove test object id, and seperate the underlyingObjectCard
