+++ apps/explorer/src/components/ownedobjects/views/DynamicFieldsCard.tsx
@@ -2,83 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useRpcClient } from '@mysten/core';
-import { getObjectDisplay } from '@mysten/sui.js';
-import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
+import { useInfiniteQuery } from '@tanstack/react-query';
 import { useRef, useCallback, useEffect } from 'react';
 
+import { UnderlyingObjectCard } from './UnderlyingObjectCard';
+
 import { SyntaxHighlighter } from '~/components/SyntaxHighlighter';
-import DisplayBox from '~/components/displaybox/DisplayBox';
 import { DisclosureBox } from '~/ui/DisclosureBox';
 import { ObjectLink } from '~/ui/InternalLink';
 import { LoadingSpinner } from '~/ui/LoadingSpinner';
 import { Tab, TabGroup, TabList, TabPanel, TabPanels } from '~/ui/Tabs';
-import { Text } from '~/ui/Text';
-import {
-    parseImageURL,
-    parseObjectType,
-    extractName,
-} from '~/utils/objectUtils';
-import { trimStdLibPrefix } from '~/utils/stringUtils';
 
 const MAX_PAGE_SIZE = 20;
-
-interface UnderlyingObjectCardProps {
-    parentId: string;
-    name: {
-        type: string;
-        value?: string;
-    };
-}
-
-function UnderlyingObjectCard({ parentId, name }: UnderlyingObjectCardProps) {
-    const rpc = useRpcClient();
-    const { data, isLoading, isError, isFetched } = useQuery(
-        ['dynamic-fields-object', parentId],
-        () =>
-            rpc.getDynamicFieldObject({
-                parentId,
-                name,
-            }),
-        {
-            enabled: !!parentId,
-        }
-    );
-
-    if (isLoading) {
-        return (
-            <div className="mt-3 pt-3">
-                <LoadingSpinner text="Loading data" />
-            </div>
-        );
-    }
-
-    if (isError || data.error || (isFetched && !data)) {
-        return null;
-    }
-    const display = getObjectDisplay(data);
-    const imgUrl = parseImageURL(display.data);
-    const objectType = parseObjectType(data);
-    const caption = extractName(display.data) || trimStdLibPrefix(objectType);
-
-    return imgUrl ? (
-        <div className="mt-3 pt-3">
-            <Text variant="body/medium" color="steel-dark">
-                Underlying Object
-            </Text>
-            <div className="item-center mt-5 flex justify-start gap-1">
-                <div className="w-16">
-                    <DisplayBox display={imgUrl} caption={caption} />
-                </div>
-                <div className="flex flex-col items-start justify-center gap-1 break-all">
-                    <ObjectLink objectId="0xf391375a604f302c61d8dbc68581592a127a4ec81f85198b560a82d521ff74e1" />
-                    <Text variant="body/medium" color="steel-dark">
-                        {caption}
-                    </Text>
-                </div>
-            </div>
-        </div>
-    ) : null;
-}
 
 export function DynamicFieldsCard({ id }: { id: string }) {
     const rpc = useRpcClient();
